Tail optimization implementation of the recursive factorial function. This will optimize the space complexity of the recursive factorial function to be O(1) instead of O(log n-1) == O(log n).
This is due to there being a parent call which will in turn call
a child.
Interpretation 1:
	The child will not make the next recursive call as a parent but
as a sibling. Thus the tree or height of the tree does not grow, only
the width of the tree grows.

Interpretation 2:
	The address from the parent call will be the same for each
recursive call. Why? The child or recursive function does not store any
values and only returns what the next function returns. Therefore, the
next child will use the same stack address as the previous child. The
stack does not grow.